# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Futbol Femeni API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alerts/user": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener alertas asignadas al usuario",
                "description": "Devuelve una lista de alertas asignadas al usuario autenticado.",
                "operationId": "9e52292218c89e986d0ac1cde073a585",
                "responses": {
                    "200": {
                        "description": "Alertas asignadas al usuario recuperadas con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alerts/unassigned": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener alertas sin llamadas asignadas",
                "description": "Devuelve las alertas que no están asociadas a ninguna llamada.",
                "operationId": "7aa6db8541652bc9317e88cffeaf1411",
                "responses": {
                    "200": {
                        "description": "Alertas sin llamadas asignadas recuperadas con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener todas las alertas",
                "description": "Devuelve una lista de todas las alertas.",
                "operationId": "4478ffd7354aa9ca6132fb9e74bdde0b",
                "responses": {
                    "200": {
                        "description": "Lista de alertas devuelta con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Crear una nueva alerta",
                "description": "Crea una nueva alerta con los datos proporcionados.",
                "operationId": "42b5297a4f4f281e294bf8339a0a80d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta creada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtener una alerta específica",
                "description": "Devuelve los datos de una alerta específica por su ID.",
                "operationId": "48e65a8f03a14c0b36563c5284e726a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta recuperada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Actualizar una alerta",
                "description": "Actualiza los datos de una alerta existente.",
                "operationId": "037adc62bb0d80e7c7a153cf2445e337",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta actualizada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Eliminar una alerta",
                "description": "Elimina una alerta específica por su ID.",
                "operationId": "6dea41ff5a0bbf10775b84b2275d50e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta eliminada con éxito."
                    },
                    "404": {
                        "description": "Alerta no encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlertStoreRequest": {
                "required": [
                    "patientId",
                    "type",
                    "startDate",
                    "isRecurring"
                ],
                "properties": {
                    "patientId": {
                        "description": "ID del paciente asociado a la alerta",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipo de alerta",
                        "type": "string",
                        "example": "Alerta médica"
                    },
                    "subType": {
                        "description": "Subtipo de alerta",
                        "type": "string",
                        "example": "Alerta de seguimiento"
                    },
                    "startDate": {
                        "description": "Fecha de inicio de la alerta, en formato ISO 8601 (Y-m-d\\TH:i)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-19T10:30"
                    },
                    "isRecurring": {
                        "description": "Indica si la alerta es recurrente",
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "description": "Descripción de la alerta",
                        "type": "string",
                        "example": "La alerta es para recordar la toma de medicamentos."
                    },
                    "recurrenceType": {
                        "description": "Tipo de recurrencia para la alerta",
                        "type": "string",
                        "example": "yearly"
                    },
                    "recurrence": {
                        "description": "Número de veces que se repetirá la alerta",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "AlertUpdateRequest": {
                "properties": {
                    "patientId": {
                        "description": "ID del paciente asociado a la alerta",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipo de alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Alerta médica"
                    },
                    "subType": {
                        "description": "Subtipo de alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Alerta de seguimiento"
                    },
                    "startDate": {
                        "description": "Fecha de inicio de la alerta, en formato ISO 8601 (Y-m-d\\TH:i)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-19T10:30"
                    },
                    "isRecurring": {
                        "description": "Indica si la alerta es recurrente",
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "description": "Descripción de la alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "La alerta es para recordar la toma de medicamentos."
                    },
                    "recurrenceType": {
                        "description": "Tipo de recurrencia para la alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "yearly"
                    },
                    "recurrence": {
                        "description": "Número de veces que se repetirá la alerta",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
            "AlertResource": {
                "description": "Respuesta de la alerta",
                "properties": {
                    "id": {
                        "description": "ID de la alerta",
                        "type": "integer"
                    },
                    "patientId": {
                        "description": "ID del paciente asociado a la alerta",
                        "type": "integer"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "type": {
                        "description": "Tipo de alerta",
                        "type": "string"
                    },
                    "subType": {
                        "description": "Subtipo de alerta",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción de la alerta",
                        "type": "string"
                    },
                    "startDate": {
                        "description": "Fecha de inicio de la alerta",
                        "type": "string",
                        "format": "date-time"
                    },
                    "isRecurring": {
                        "description": "Indica si la alerta es recurrente",
                        "type": "boolean"
                    },
                    "recurrenceType": {
                        "description": "Tipo de recurrencia de la alerta",
                        "type": "string"
                    },
                    "recurrence": {
                        "description": "Número de recurrencias",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "description": "Recurso de paciente",
                "properties": {
                    "id": {
                        "description": "ID del paciente",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del paciente",
                        "type": "string"
                    },
                    "lastName": {
                        "description": "Apellido del paciente",
                        "type": "string"
                    },
                    "birthDate": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date"
                    },
                    "userId": {
                        "description": "ID del usuario asociado al paciente",
                        "type": "integer"
                    },
                    "addressStreet": {
                        "description": "Calle de la dirección del paciente",
                        "type": "string"
                    },
                    "addressNumber": {
                        "description": "Número de la dirección del paciente",
                        "type": "integer"
                    },
                    "addressFloor": {
                        "description": "Piso de la dirección del paciente",
                        "type": "string"
                    },
                    "addressDoor": {
                        "description": "Puerta de la dirección del paciente",
                        "type": "string"
                    },
                    "addressPostalCode": {
                        "description": "Código postal de la dirección del paciente",
                        "type": "integer"
                    },
                    "addressCity": {
                        "description": "Ciudad de la dirección del paciente",
                        "type": "string"
                    },
                    "addressProvince": {
                        "description": "Provincia de la dirección del paciente",
                        "type": "string"
                    },
                    "addressCountry": {
                        "description": "País de la dirección del paciente",
                        "type": "string"
                    },
                    "dni": {
                        "description": "DNI del paciente",
                        "type": "string"
                    },
                    "healthCardNumber": {
                        "description": "Número de la tarjeta sanitaria del paciente",
                        "type": "string"
                    },
                    "prefix": {
                        "description": "Prefijo del teléfono del paciente",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Número de teléfono del paciente",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email"
                    },
                    "language": {
                        "description": "Idioma preferido del paciente",
                        "type": "string"
                    },
                    "zoneId": {
                        "description": "ID de la zona asignada al paciente",
                        "type": "integer"
                    },
                    "healthSituation": {
                        "description": "Situación de salud del paciente",
                        "type": "string"
                    },
                    "situationPersonalFamily": {
                        "description": "Situación personal/familiar del paciente",
                        "type": "string"
                    },
                    "housingSituationType": {
                        "description": "Tipo de situación de vivienda del paciente",
                        "type": "string"
                    },
                    "housingSituationStatus": {
                        "description": "Estado de situación de vivienda del paciente",
                        "type": "string"
                    },
                    "housingSituationNumberOfRooms": {
                        "description": "Número de habitaciones en la vivienda del paciente",
                        "type": "integer"
                    },
                    "housingSituationLocation": {
                        "description": "Ubicación de la vivienda del paciente",
                        "type": "string"
                    },
                    "personalAutonomy": {
                        "description": "Autonomía personal del paciente",
                        "type": "string"
                    },
                    "economicSituation": {
                        "description": "Situación económica del paciente",
                        "type": "string"
                    },
                    "contacts": {
                        "description": "Contactos del paciente",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}